Move history classes into separate app
