+++ api/history/exceptions.py
@@ -1,0 +1,6 @@
+class HistoryFactoryNotFound(Exception):
+    pass
+
+
+class HistoryItemNotFound(Exception):
+    pass
