+++ api/history/factories/__init__.py
@@ -1,5 +1,6 @@
 from .assessments import AssessmentHistoryFactory
 from .barriers import BarrierHistoryFactory
+from .core import HistoryItemFactory
 from .notes import NoteHistoryFactory
 from .public_barriers import PublicBarrierHistoryFactory
 from .public_barrier_notes import PublicBarrierNoteHistoryFactory
