+++ api/history/factories/team_members.py
@@ -1,17 +1,6 @@
 from api.collaboration.models import TeamMember
-from .base import BaseHistoryItem, HistoryItemFactoryBase
-
-
-class TeamMemberHistoryItem(BaseHistoryItem):
-    model = "team_member"
-    field = "user"
-
-    def get_value(self, record):
-        if record and not record.archived:
-            return {
-                "user": self._format_user(record.user),
-                "role": record.role,
-            }
+from .base import HistoryItemFactoryBase
+from ..items.team_members import TeamMemberHistoryItem
 
 
 class TeamMemberHistoryFactory(HistoryItemFactoryBase):
