+++ api/history/items/barriers.py
@@ -1,6 +1,7 @@
 from api.barriers.models import BarrierInstance
 from api.metadata.utils import get_country
-from .base import BaseHistoryItem, HistoryItemFactoryBase
+from .base import BaseHistoryItem
+from ..factories.base import HistoryItemFactoryBase
 
 
 class BaseBarrierHistoryItem(BaseHistoryItem):
@@ -152,36 +153,3 @@ class TitleHistoryItem(BaseBarrierHistoryItem):
 
 class TradeDirectionHistoryItem(BaseBarrierHistoryItem):
     field = "trade_direction"
-
-
-class BarrierHistoryFactory(HistoryItemFactoryBase):
-    """
-    Polymorphic wrapper for barrier HistoryItem classes
-    """
-
-    class_lookup = {}
-    history_item_classes = (
-        ArchivedHistoryItem,
-        CategoriesHistoryItem,
-        CommoditiesHistoryItem,
-        CompaniesHistoryItem,
-        EndDateHistoryItem,
-        IsSummarySensitiveHistoryItem,
-        LocationHistoryItem,
-        PriorityHistoryItem,
-        ProductHistoryItem,
-        PublicEligibilitySummaryHistoryItem,
-        ScopeHistoryItem,
-        SectorsHistoryItem,
-        SourceHistoryItem,
-        StatusHistoryItem,
-        SummaryHistoryItem,
-        TagsHistoryItem,
-        TitleHistoryItem,
-        TradeDirectionHistoryItem,
-    )
-    history_types = ("~", "+")
-
-    @classmethod
-    def get_history(cls, barrier_id):
-        return BarrierInstance.history.filter(id=barrier_id).order_by("history_date")
