+++ api/history/items/team_members.py
@@ -1,0 +1,15 @@
+from api.collaboration.models import TeamMember
+from .base import BaseHistoryItem
+from ..factories.base import HistoryItemFactoryBase
+
+
+class TeamMemberHistoryItem(BaseHistoryItem):
+    model = "team_member"
+    field = "user"
+
+    def get_value(self, record):
+        if record and not record.archived:
+            return {
+                "user": self._format_user(record.user),
+                "role": record.role,
+            }
